Constraints and States read

+++ app/src/main/java/com/sg2022/we_got_the_moves/db/entity/daos/ConstraintDao.java
@@ -33,5 +33,5 @@ public interface ConstraintDao {
   Single<List<Constraint>> getAllSingle();
 
   @Query("SELECT * FROM [Constraint] WHERE [Constraint].id == :constraintId")
-  Single<Constraint> getSingle(int constraintId);
+  Single<Constraint> getSingle(Long constraintId);
 }
